rename and relocate log
