+++ src/globalvars.ts
@@ -1,0 +1,9 @@
+export const log = {
+  initialEmbezzlersFought: 0,
+  digitizedEmbezzlersFought: 0,
+};
+
+export const globalOptions: { ascending: boolean; stopTurncount: number | null } = {
+  stopTurncount: null,
+  ascending: false,
+};
